pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "autoflake>=2.3.1",
    "black>=24.10.0",
    "fastapi[standard]>=0.115.7",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "pydantic>=2.10.5",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.37",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]
